+++ packages/plugin-event-property-attribution-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3-SR-4497-custom-transport-hooks.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-event-property-attribution-browser@0.2.2...@amplitude/plugin-event-property-attribution-browser@0.2.3-SR-4497-custom-transport-hooks.0) (2026-06-09)
+
+**Note:** Version bump only for package @amplitude/plugin-event-property-attribution-browser
+
+
+
+
+
 ## [0.2.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-event-property-attribution-browser@0.2.1...@amplitude/plugin-event-property-attribution-browser@0.2.2) (2026-06-05)
 
 **Note:** Version bump only for package @amplitude/plugin-event-property-attribution-browser